Move generated pb files from internal/pb to pkg/pb (#502)
* Move generated pb files from internal/pb to pkg/pb

* Update base on feedback
