+++ api/evaluator.proto
@@ -14,7 +14,7 @@
 
 syntax = "proto3";
 package api;
-option go_package = "internal/pb";
+option go_package = "pkg/pb";
 
 import "api/messages.proto";
 import "google/api/annotations.proto";
@@ -56,12 +56,12 @@ option (grpc.gateway.protoc_gen_swagger.options.openapiv2_swagger) = {
 
 message EvaluateRequest {
   // List of Matches to evaluate.
-  repeated Match match = 1;
+  repeated api.Match match = 1;
 }
 
 message EvaluateResponse {
   // Accepted list of Matches.
-  repeated Match match = 1;
+  repeated api.Match match = 1;
 }
 
 // The service implementing the Evaluator API that is called to evaluate
