+++ cloudbuild.yaml
@@ -93,6 +93,7 @@ steps:
   - name: 'go-vol'
     path: '/go'
   waitFor: ['Build: Assets']
+
 - id: 'Test: Services'
   name: 'gcr.io/$PROJECT_ID/open-match-build'
   args: ['make', 'GOPROXY=off', 'GOLANG_TEST_COUNT=25', 'test']
@@ -100,6 +101,7 @@ steps:
   - name: 'go-vol'
     path: '/go'
   waitFor: ['Build: Assets']
+
 - id: 'Build: Docker Images'
   name: 'gcr.io/$PROJECT_ID/open-match-build'
   args: ['make', 'VERSION_SUFFIX=$SHORT_SHA', 'push-images', '-j8']
@@ -156,10 +158,6 @@ artifacts:
             - cmd/synchronizer/synchronizer
             - cmd/minimatch/minimatch
             - cmd/swaggerui/swaggerui
-            - examples/functions/golang/soloduel/soloduel
-            - examples/evaluator/golang/simple/simple
-            - tools/certgen/certgen
-            - tools/reaper/reaper
             - install/yaml/install.yaml
             - install/yaml/install-demo.yaml
             - install/yaml/01-redis-chart.yaml
@@ -167,6 +165,11 @@ artifacts:
             - install/yaml/03-prometheus-chart.yaml
             - install/yaml/04-grafana-chart.yaml
             - install/yaml/05-jaeger-chart.yaml
+            - examples/functions/golang/soloduel/soloduel
+            - examples/evaluator/golang/simple/simple
+            - tools/certgen/certgen
+            - tools/reaper/reaper
+
 images:
 - 'gcr.io/$PROJECT_ID/openmatch-backend:${_OM_VERSION}-${SHORT_SHA}'
 - 'gcr.io/$PROJECT_ID/openmatch-frontend:${_OM_VERSION}-${SHORT_SHA}'
