+++ examples/functions/golang/pool/matchfunction.go
@@ -21,8 +21,8 @@ package pool
 
 import (
 	"github.com/rs/xid"
-	"open-match.dev/open-match/internal/pb"
 	mmfHarness "open-match.dev/open-match/pkg/harness/function/golang"
+	"open-match.dev/open-match/pkg/pb"
 )
 
 var (
@@ -45,7 +45,7 @@ func MakeMatches(params *mmfHarness.MatchFunctionParams) []*pb.Match {
 		result = append(result, &pb.Match{
 			MatchId:       xid.New().String(),
 			MatchProfile:  params.ProfileName,
-			MatchFunction: "pool-based-match",
+			MatchFunction: matchName,
 			Ticket:        tickets,
 			Roster:        []*pb.Roster{roster},
 		})
