+++ internal/app/mmlogic/mmlogic_service.go
@@ -21,7 +21,8 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"open-match.dev/open-match/internal/config"
-	"open-match.dev/open-match/internal/pb"
+	"open-match.dev/open-match/pkg/pb"
+
 	"open-match.dev/open-match/internal/statestore"
 )
