+++ internal/rpc/insecure_test.go
@@ -20,7 +20,7 @@ import (
 	"testing"
 	"time"
 
-	"open-match.dev/open-match/internal/pb"
+	"open-match.dev/open-match/pkg/pb"
 
 	"github.com/stretchr/testify/assert"
 	"google.golang.org/grpc"
