+++ internal/rpc/server_test.go
@@ -23,7 +23,7 @@ import (
 	"testing"
 	"time"
 
-	"open-match.dev/open-match/internal/pb"
+	"open-match.dev/open-match/pkg/pb"
 
 	"github.com/stretchr/testify/assert"
 	"google.golang.org/grpc"
