+++ internal/statestore/redis.go
@@ -26,8 +26,8 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"open-match.dev/open-match/internal/config"
-	"open-match.dev/open-match/internal/pb"
 	"open-match.dev/open-match/internal/set"
+	"open-match.dev/open-match/pkg/pb"
 )
 
 var (
