+++ internal/testing/fake_frontend.go
@@ -20,7 +20,7 @@ import (
 
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-	"open-match.dev/open-match/internal/pb"
+	"open-match.dev/open-match/pkg/pb"
 )
 
 // FakeFrontend is an empty gRPC handler.
