+++ pkg/harness/function/golang/doc.go
@@ -18,5 +18,5 @@
 // Note that the main package for the harness does very little except read the
 // config and set up logging and metrics, then start the server. The harness
 // functionality is implemented in harness.go, which implements the gRPC server
-// defined in the internal/pb/matchfunction.pb.go file.
+// defined in the pkg/pb/matchfunction.pb.go file.
 package golang
