+++ test/e2e/minimatch/evaluator_setup.go
