+++ test/e2e/minimatch/match_test.go
@@ -22,8 +22,8 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
-	"open-match.dev/open-match/internal/pb"
 	rpcTesting "open-match.dev/open-match/internal/rpc/testing"
+	"open-match.dev/open-match/pkg/pb"
 )
 
 func TestFetchMatches(t *testing.T) {
