+++ test/e2e/minimatch/ticket_test.go
@@ -23,7 +23,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-	"open-match.dev/open-match/internal/pb"
+	"open-match.dev/open-match/pkg/pb"
 
 	rpcTesting "open-match.dev/open-match/internal/rpc/testing"
 )
